tests pkg/sandbox/teardown
