+++ cmd/sandbox/teardown_test.go
@@ -1,8 +1,6 @@
 package sandbox
 
 import (
-	"context"
-	"fmt"
 	"testing"
 
 	"github.com/docker/docker/api/types"
@@ -12,53 +10,13 @@ import (
 	"github.com/flyteorg/flytectl/pkg/docker/mocks"
 	"github.com/flyteorg/flytectl/pkg/k8s"
 	k8sMocks "github.com/flyteorg/flytectl/pkg/k8s/mocks"
-	"github.com/flyteorg/flytectl/pkg/sandbox"
 	"github.com/flyteorg/flytectl/pkg/util"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
 
-var containers []types.Container
-
-func TestTearDownFunc(t *testing.T) {
-	container1 := types.Container{
-		ID: "FlyteSandboxClusterName",
-		Names: []string{
-			docker.FlyteSandboxClusterName,
-		},
-	}
-	containers = append(containers, container1)
-
-	t.Run("Success", func(t *testing.T) {
-		ctx := context.Background()
-		mockDocker := &mocks.Docker{}
-		mockDocker.OnContainerList(ctx, types.ContainerListOptions{All: true}).Return(containers, nil)
-		mockDocker.OnContainerRemove(ctx, mock.Anything, types.ContainerRemoveOptions{Force: true}).Return(nil)
-		mockK8sContextMgr := &k8sMocks.ContextOps{}
-		k8s.ContextMgr = mockK8sContextMgr
-		mockK8sContextMgr.OnRemoveContextMatch(mock.Anything).Return(nil)
-		err := sandbox.Teardown(ctx, mockDocker)
-		assert.Nil(t, err)
-	})
-	t.Run("Error", func(t *testing.T) {
-		ctx := context.Background()
-		mockDocker := &mocks.Docker{}
-		mockDocker.OnContainerList(ctx, types.ContainerListOptions{All: true}).Return(containers, nil)
-		mockDocker.OnContainerRemove(ctx, mock.Anything, types.ContainerRemoveOptions{Force: true}).Return(fmt.Errorf("err"))
-		err := sandbox.Teardown(ctx, mockDocker)
-		assert.NotNil(t, err)
-	})
-	t.Run("Error", func(t *testing.T) {
-		ctx := context.Background()
-		mockDocker := &mocks.Docker{}
-		mockDocker.OnContainerList(ctx, types.ContainerListOptions{All: true}).Return(nil, fmt.Errorf("err"))
-		err := sandbox.Teardown(ctx, mockDocker)
-		assert.NotNil(t, err)
-	})
-
-}
-
 func TestTearDownClusterFunc(t *testing.T) {
+	var containers []types.Container
 	_ = util.SetupFlyteDir()
 	_ = util.WriteIntoFile([]byte("data"), configutil.FlytectlConfig)
 	s := testutils.Setup()
@@ -66,6 +24,10 @@ func TestTearDownClusterFunc(t *testing.T) {
 	mockDocker := &mocks.Docker{}
 	mockDocker.OnContainerList(ctx, types.ContainerListOptions{All: true}).Return(containers, nil)
 	mockDocker.OnContainerRemove(ctx, mock.Anything, types.ContainerRemoveOptions{Force: true}).Return(nil)
+	mockK8sContextMgr := &k8sMocks.ContextOps{}
+	mockK8sContextMgr.OnRemoveContext(mock.Anything).Return(nil)
+	k8s.ContextMgr = mockK8sContextMgr
+
 	docker.Client = mockDocker
 	err := teardownSandboxCluster(ctx, []string{}, s.CmdCtx)
 	assert.Nil(t, err)
